+++ src/main.js
@@ -1,7 +1,7 @@
 import { Viewer } from "@xeokit/xeokit-sdk";
-import "../static/js/error-handler.js";
-import "../static/js/upload-handler.js";
-import "../static/js/design-insights.js";
+import "./error-handler.js";
+import "./upload-handler.js";
+import "./design-insights.js";
 import { initViewer, loadXeokitSDK } from "./xeokit_viewer.js";
 import { jsPDF } from "jspdf";
